requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.9.1
werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
blinker==1.7.0
